+++ src/CostFunction.h
@@ -11,19 +11,19 @@ public:
   static std::unique_ptr<CostFunction> instance(options::CostFunctionType type);
 
   virtual ~CostFunction() = default;
-  
-  virtual double operator()(double value, double target) const = 0;
-  virtual double derivative(double value, double target) const = 0;
+
+  virtual float operator()(float value, float target) const = 0;
+  virtual float derivative(float value, float target) const = 0;
 };
 
 class QuadraticCostFunction : public CostFunction {
 public:
-  double operator()(double value, double target) const override;
-  double derivative(double value, double target) const override;
+  float operator()(float value, float target) const override;
+  float derivative(float value, float target) const override;
 };
 
 class CostEntropyCostFunction : public CostFunction {
 public:
-  double operator()(double value, double target) const override;
-  double derivative(double value, double target) const override;
+  float operator()(float value, float target) const override;
+  float derivative(float value, float target) const override;
 };
