+++ src/Layer.h
@@ -22,36 +22,36 @@ public:
 
   std::size_t numberOfNeurons() const;
 
-  double loss() const;
+  float loss() const;
 
-  const Eigen::VectorXd &outputs() const;
+  const Eigen::VectorXf &outputs() const;
 
-  const Eigen::VectorXd &errors() const;
+  const Eigen::VectorXf &errors() const;
 
-  const Eigen::MatrixXd &weights() const;
+  const Eigen::MatrixXf &weights() const;
 
-  Eigen::MatrixXd computeGradients() const;
+  Eigen::MatrixXf computeGradients() const;
 
-  void updateOutputs(const Eigen::VectorXd &inputs);
+  void updateOutputs(const Eigen::VectorXf &inputs);
 
   void updateErrors(const Layer &nextLayer);
 
-  void updateErrors(const Eigen::VectorXd &targets,
+  void updateErrors(const Eigen::VectorXf &targets,
                     const CostFunction &costFunction);
 
-  void updateWeights(double learnRate);
+  void updateWeights(float learnRate);
 
-  void updateWeights(const Eigen::MatrixXd &gradients, double learnRate);
+  void updateWeights(const Eigen::MatrixXf &gradients, float learnRate);
 
 private:
   const std::size_t m_id;
   const std::size_t m_numberOfInputs;
   const std::size_t m_numberOfNeurons;
   std::unique_ptr<ActivationFunction> m_activationFunction;
-  Eigen::VectorXd m_intermediateQtys;
-  Eigen::VectorXd m_inputs;
-  Eigen::VectorXd m_outputs;
-  Eigen::VectorXd m_errors;
-  Eigen::MatrixXd m_weights;
-  double m_loss;
+  Eigen::VectorXf m_intermediateQtys;
+  Eigen::VectorXf m_inputs;
+  Eigen::VectorXf m_outputs;
+  Eigen::VectorXf m_errors;
+  Eigen::MatrixXf m_weights;
+  float m_loss;
 };
