+++ src/OptimizationAlgorithm.h
@@ -22,30 +22,30 @@ public:
 
   virtual ~OptimizationAlgorithm() = default;
 
-  virtual void run(TrainingBatch batch, std::size_t maxEpoch, double learnRate,
-                   double lossGoal);
+  virtual void run(TrainingBatch batch, std::size_t maxEpoch, float learnRate,
+                   float lossGoal);
 
   std::size_t epochsCount() const;
 
-  double loss() const;
+  float loss() const;
 
 protected:
   OptimizationAlgorithm(options::CostFunctionType costFunction,
                         std::vector<Layer> &layers);
 
-  virtual void beforeRun(double learnRate);
+  virtual void beforeRun(float learnRate);
   virtual void afterSample();
   virtual void afterEpoch();
 
   void updateLoss();
-  void forwardPropagate(const std::vector<double> &inputs);
-  void backwardPropagate(const std::vector<double> &outputs);
+  void forwardPropagate(const std::vector<float> &inputs);
+  void backwardPropagate(const std::vector<float> &outputs);
   void preprocess(TrainingBatch &batch) const;
 
   std::vector<Layer> &m_layers;
   std::unique_ptr<CostFunction> m_costFunction;
   std::size_t m_epochsCount;
   std::size_t m_samplesCount;
-  double m_learnRate;
-  double m_loss;
+  float m_learnRate;
+  float m_loss;
 };
